Widen conntrack extension offsets to u16 and extend the ML feature vector
struct nf_ct_ext stored each extension's offset and the block length in
u8, capping the whole extension area at 255 bytes. Change offset[] and len
to u16 and raise the nf_ct_ext_add() compile-time check from 255 to 65535,
so the extension block can be larger; the check still keeps every offset
and the length within the field width.

struct nf_conn_ml now carries the per-direction payload byte totals
(bytes_fwd, bytes_bwd) and a payload sample count (pktlen_count), and its
inter-arrival sum is named iat_sum_us to reflect that it is accumulated in
microseconds. The packet-length fields now describe the L4 payload only.
The vector is 144 bytes.

+++ include/net/netfilter/nf_conntrack_extend.h
@@ -35,10 +35,16 @@ enum nf_ct_ext_id {
 	NF_CT_EXT_NUM,
 };
 
-/* Extensions: optional stuff which isn't permanently in struct. */
+/* Extensions: optional stuff which isn't permanently in struct.
+ *
+ * offset[] holds each extension's byte offset within data[]; len holds the
+ * total size of the block. Both are u16, and nf_ct_ext_add() enforces
+ * total_extension_size() <= U16_MAX at compile time (BUILD_BUG_ON), so every
+ * offset and the length always fit these fields.
+ */
 struct nf_ct_ext {
-	u8 offset[NF_CT_EXT_NUM];
-	u8 len;
+	u16 offset[NF_CT_EXT_NUM];
+	u16 len;
 	unsigned int gen_id;
 	char data[] __aligned(8);
 };
